### 2025-04-30 #8143 @editorialbot
[PRE REVIEW]: dfreproject: A Python package for astronomical reprojection
**Submitting author:** @crhea93 (<a href="http://orcid.org/0000-0003-2001-1076">Carter Rhea</a>)
**Repository:** https://github.com/DragonflyTelescope/dfreproject
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
