### 2025-04-30 #8143 @warrickball
Hi @crhea93, thanks for your submission and sorry for the delay in starting our pre-review process! All the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review until an editor with the relevant expertise is available to take it on. The AASS track at JOSS currently has a significant backlog, so this might take several months. Thanks for your patience!

In the meantime, you can probably fix the "Statement of need" of warning by removing the numbering from your section titles. You should also try to reduce the paper to closer to our [1000 word limit](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain), though this isn't a strict limit. In your case, I think most (if not all) of Section 3 can be removed and placed in the online documentation as reference.

To render your changes to the paper here, post a message that starts with `@editorialbot generate pdf`.
