### 2025-07-01 DragonflyTelescope/dfreproject#18 @keflavich
Documentation / JOSS review: Clarify usage in paper example
Related to: https://github.com/openjournals/joss-reviews/issues/8525

In the example in the paper, `calculate_reprojection` is shown with no argument specifying threads and/or GPU.

Since the emphasis of the paper is on performance, please expand the example to show how to specify whether a GPU is to be used (with `device=` in the call) and/or with threads.

Relatedly, please clarify whether the CPU demo was single- or multi-threaded.  It presumably is single, but noting this and emphasizing that multithreading is supported helps show users the unique capabilities of this reprojection tool.
