+++ _sources/autoapi/dfreproject/reproject/index.rst.txt
@@ -98,7 +98,12 @@ Module Contents
    :type interpolation_mode: str
 
 
-.. py:class:: Reproject(source_hdus: List[astropy.io.fits.PrimaryHDU], target_wcs: astropy.wcs.WCS, shape_out: Tuple[int, int], device: str = None, num_threads: int = None)
+.. py:class:: Reproject(source_hdus: List[astropy.io.fits.PrimaryHDU], target_wcs: astropy.wcs.WCS, shape_out: Tuple[int, int], device: str = None, num_threads: int = None, requires_grad: bool = False)
+
+   .. py:attribute:: requires_grad
+      :value: False
+
+
 
    .. py:attribute:: batch_source_images
 
@@ -188,25 +193,26 @@ Module Contents
 
 
 
-.. py:function:: calculate_reprojection(source_hdus: Union[astropy.io.fits.PrimaryHDU, Tuple[numpy.ndarray, Union[astropy.wcs.WCS, astropy.io.fits.Header]], List[Union[astropy.io.fits.PrimaryHDU, Tuple[numpy.ndarray, Union[astropy.wcs.WCS, astropy.io.fits.Header]]]]], target_wcs: Union[astropy.wcs.WCS, astropy.io.fits.Header], shape_out: Optional[Tuple[int, int]] = None, order: str = 'nearest', device: str = None, num_threads: int = None)
+.. py:function:: calculate_reprojection(source_hdus: Union[astropy.io.fits.PrimaryHDU, dfreproject.tensorhdu.TensorHDU, Tuple[numpy.ndarray, Union[astropy.wcs.WCS, astropy.io.fits.Header]], Tuple[torch.Tensor, Union[astropy.wcs.WCS, astropy.io.fits.Header]], List[Union[astropy.io.fits.PrimaryHDU, Tuple[numpy.ndarray, Union[astropy.wcs.WCS, astropy.io.fits.Header]]]]], target_wcs: Union[astropy.wcs.WCS, astropy.io.fits.Header], shape_out: Optional[Tuple[int, int]] = None, order: str = 'nearest', device: str = None, num_threads: int = None, requires_grad: bool = False)
 
    Reproject an astronomical image from a source WCS to a target WCS.
 
-   This high-level function provides a convenient interface for image dfreproject,
+   This high-level function provides a convenient interface for image reprojection,
    handling all the necessary steps: WCS extraction, tensor creation, and interpolation.
-   It converts FITS HDU objects to the internal representation, performs the dfreproject,
-   and returns the resulting image as a PyTorch tensor.
+   It converts FITS HDU objects to the internal representation, performs the reprojection,
+   and returns the resulting image as a NumPy array or PyTorch tensor.
 
    :param source_hdus: PrimaryHDU,
+                       TensorHDU,
                        Tuple[np.ndarray, Union[WCS, Header]],
+                       Tuple[torch.Tensor, Union[WCS, Header]],
                        List[Union[PrimaryHDU, Tuple[np.ndarray, Union[WCS, Header]]]]
    :type source_hdus: Union[
-   :param ]: The source image HDU list containing the image data to be reprojected and
-             its associated WCS information in the header. The user can also pass a single instance or a tuple
-             of the following form: (source_data, source_wcs).
-   :param target_wcs: WCS information fopr the target. If a Header is passed it will be transformed to WCS.
+   :param ]: The source image(s) to be reprojected, which can be a PrimaryHDU, TensorHDU,
+             a tuple of (data, WCS/Header), or a list of such items.
+   :param target_wcs: WCS information for the target. If a Header is passed it will be converted to WCS.
    :type target_wcs: Union[WCS, Header]
-   :param shape_out: Shape of the resampled array. If no argument is passed, then the output shape will be the same as the input shape.
[…]
