### 2025-07-11 DragonflyTelescope/dfreproject#18 @crhea93
@keflavich Thank you very much for pointing this out! I've updated the paper to reflect these comments.
