### 2025-07-14 DragonflyTelescope/dfreproject#24 @keflavich
Documentation: missing test instructions
At least, I didn't find them by searching.

I was able to run tests by running:
```bash
pytest tests
```

in the root directory.  This should be documented
