### 2025-07-14 #8525 @keflavich
https://github.com/DragonflyTelescope/dfreproject/issues/24 is minor - tests exists, but running them isn't explicitly documented
