### 2025-07-16 DragonflyTelescope/dfreproject#24 @crhea93
Addressed this in a recent commit.

In the ReadMe we now have this section
