+++ _sources/autoapi/dfreproject/reproject/index.rst.txt
@@ -51,7 +51,8 @@ Module Contents
 
    Function to validate the requested interpolation order.
 
-   The order must be one of the following: {valid_orders}
+   The order must be one of the following: "bicubic", "bilinear",
+   "nearest-neighbors". "nearest" is an alias for "nearest-neighbors".
 
    :param order: Interpolation order to validate.
    :type order: str
@@ -64,41 +65,41 @@ Module Contents
 
 .. py:function:: atan2d(y: torch.Tensor, x: torch.Tensor) -> torch.Tensor
 
-   PyTorch implementation of WCSLib's atan2d function
+   PyTorch implementation of WCSLib's atan2d function.
 
-   :param y: y coordinate(s)
+   :param y: y coordinate(s).
    :type y: torch.Tensor
-   :param x: x coordinate(s)
+   :param x: x coordinate(s).
    :type x: torch.Tensor
 
-   :returns: atan2d(y, x) in degrees
+   :returns: atan2d(y, x) in degrees.
    :rtype: torch.Tensor
 
 
 .. py:function:: sincosd(angle_deg: torch.Tensor) -> torch.Tensor
 
-   PyTorch implementation of WCSLib's sincosd function
+   PyTorch implementation of WCSLib's sincosd function.
 
-   :param angle_deg: angle in degrees
+   :param angle_deg: angle in degrees.
    :type angle_deg: torch.Tensor
 
-   :returns: sin(angle) in degrees, cos(angle) in degrees
+   :returns: sin(angle) in degrees, cos(angle) in degrees.
    :rtype: tuple(torch.Tensor, torch.Tensor)
 
 
 .. py:function:: interpolate_image(source_image: torch.Tensor, grid: torch.Tensor, interpolation_mode: str) -> torch.Tensor
 
-   JIT-compiled image interpolation using grid_sample
+   JIT-compiled image interpolation using grid_sample.
 
-   :param source_image: Source image to interpolate
+   :param source_image: Source image to interpolate.
    :type source_image: torch.Tensor
-   :param grid: Grid on which to interpolate
+   :param grid: Grid on which to interpolate.
    :type grid: torch.Tensor
-   :param interpolation_mode: Interpolation mode to use
+   :param interpolation_mode: Interpolation mode to use.
    :type interpolation_mode: str
 
 
-.. py:class:: Reproject(source_hdus: List[astropy.io.fits.PrimaryHDU], target_wcs: astropy.wcs.WCS, shape_out: Tuple[int, int], device: str = None, num_threads: int = None, requires_grad: bool = False)
+.. py:class:: Reproject(source_hdus: List[astropy.io.fits.PrimaryHDU], target_wcs: astropy.wcs.WCS, shape_out: Tuple[int, int], device: str = None, num_threads: int = None, requires_grad: bool = False, compute_jacobian: bool = True)
 
    .. py:attribute:: requires_grad
       :value: False
@@ -114,26 +115,35 @@ Module Contents
    .. py:attribute:: target_wcs_params
 
 
+   .. py:attribute:: target_wcs
+
+
    .. py:attribute:: target_grid
+
+
+   .. py:attribute:: compute_jacobian
+      :value: True
+
 
 
    .. py:method:: calculate_skyCoords(x_grid=None, y_grid=None) -> Tuple[torch.Tensor, torch.Tensor]
 
       Calculate sky coordinates.
+
       There are four primary steps:
       1. Apply shift
       2. Apply SIP distortion
       3. Apply CD matrix
       4. Apply transformation to celestial coordinates uing the Gnomonic Projection
 
-      These steps use the target wcs parameters
+      These steps use the target wcs parameters.
 
[…]
