+++ _sources/autoapi/dfreproject/reproject/index.rst.txt
@@ -99,7 +99,7 @@ Module Contents
    :type interpolation_mode: str
 
 
-.. py:class:: Reproject(source_hdus: List[astropy.io.fits.PrimaryHDU], target_wcs: astropy.wcs.WCS, shape_out: Tuple[int, int], device: str = None, num_threads: int = None, requires_grad: bool = False, compute_jacobian: bool = True)
+.. py:class:: Reproject(source_hdus: List[astropy.io.fits.PrimaryHDU], target_wcs: astropy.wcs.WCS, shape_out: Tuple[int, int], device: str = None, num_threads: int = None, requires_grad: bool = False, conserve_flux: bool = True, compute_jacobian: bool = True)
 
    .. py:attribute:: requires_grad
       :value: False
@@ -119,6 +119,11 @@ Module Contents
 
 
    .. py:attribute:: target_grid
+
+
+   .. py:attribute:: conserve_flux
+      :value: True
+
 
 
    .. py:attribute:: compute_jacobian
@@ -219,7 +224,7 @@ Module Contents
 
 
 
-.. py:function:: calculate_reprojection(source_hdus: Union[astropy.io.fits.PrimaryHDU, dfreproject.tensorhdu.TensorHDU, Tuple[numpy.ndarray, Union[astropy.wcs.WCS, astropy.io.fits.Header]], Tuple[torch.Tensor, Union[astropy.wcs.WCS, astropy.io.fits.Header]], List[Union[astropy.io.fits.PrimaryHDU, Tuple[numpy.ndarray, Union[astropy.wcs.WCS, astropy.io.fits.Header]]]]], target_wcs: Union[astropy.wcs.WCS, astropy.io.fits.Header], shape_out: Optional[Tuple[int, int]] = None, order: str = 'nearest', device: str = None, num_threads: int = None, requires_grad: bool = False, compute_jacobian: bool = True)
+.. py:function:: calculate_reprojection(source_hdus: Union[astropy.io.fits.PrimaryHDU, dfreproject.tensorhdu.TensorHDU, Tuple[numpy.ndarray, Union[astropy.wcs.WCS, astropy.io.fits.Header]], Tuple[torch.Tensor, Union[astropy.wcs.WCS, astropy.io.fits.Header]], List[Union[astropy.io.fits.PrimaryHDU, Tuple[numpy.ndarray, Union[astropy.wcs.WCS, astropy.io.fits.Header]]]]], target_wcs: Union[astropy.wcs.WCS, astropy.io.fits.Header], shape_out: Optional[Tuple[int, int]] = None, order: str = 'nearest', device: str = None, num_threads: int = None, requires_grad: bool = False, conserve_flux: bool = True, compute_jacobian: bool = True)
 
    Reproject an astronomical image from a source WCS to a target WCS.
 
@@ -252,7 +257,10 @@ Module Contents
    :type num_threads: int, optional
    :param requires_grad: If True, enables autograd for PyTorch tensors.
    :type requires_grad: bool, optional
-   :param compute_jacobian: If True, enables flux conservation through Jacobian calculation. Note
+   :param conserve_flux: If True, enables flux conservation through footprint calculations.
+                         By default, this is set to True.
+   :type conserve_flux: bool, optional
+   :param compute_jacobian: If True, enables non-linear flux conservation through Jacobian calculation. Note
                             that this slightly increases RAM usage.
                             By default, this is set to True.
                             If there is no SIP distortion, users can set this to False.
