### 2025-09-04 DragonflyTelescope/dfreproject#27 @svank
Flux conservation seems incomplete
Hi @crhea93, I'm really glad to see you've gotten gpu-accelerated reprojection working! I think overall the package is looking pretty good! I spent some time digging into the flux conservation, since that's been important in my work, and I think it's not quite complete yet (though it might currently be enough for what Dragonfly's doing).

First, just to check, after #26, with `conserve_flux=True, compute_jacobian=True`, assuming the content of the source image fits entirely inside the output frame, is the flux conservation intended to ensure the sum of the input image equals the sum the output image? If so, it doesn't seem to be achieving that (at least not always), but I think the fix is simple.

In `compute_sip_jacobian`, it appears to compute the Jacobian of the transformation from pixel coordinates to what the distortion paper calls intermediate pixel coordinates. It computes the derivative of the SIP transformation and then adds in the PC matrix, and then the Jacobian is d(intermediate pixel coordinate) / d(raw pixel coordinate). But to achieve flux conservation, the Jacobian needs to measure d(source pixel coordinate) / d(output pixel coordinate)---at least, that's how it is for adaptive reprojection in the astropy reproject package. Critically, that means the Jacobian includes the effects of both the input and output frames' WCSes.

To test this, I'm using the current main branch, commit `03c746dab0de428e589359b70b982a42d9a23b20`. I copied the setup in `test_flux_conservation` in `test_interpolation.py`, with de-fixture-ified versions of `setup_real_projection`, `source_fits_file`, etc. In `sip_wcs`, I reduced the SIP degree to 2, scaled up the amplitude of the SIP coefficients, and increased CDELT for the output frame.([code here](https://gist.github.com/svank/0ce278449f8258c60e012ff074a81ec6)) That gives a reprojection that looks like this:

With the code as-is, the two images sum to 314154 and 78904, respectively. After hacking up `Reproject.interpolate_source_image` to compute d(source pixel coordinate) / d(output pixel coordinate), the two images sum to 314154 and 314170, a difference of 0.005%.

To do this, I made this change, which is definitely not the cleanest way to do this:
```
diff --git a/src/dfreproject/reproject.py b/src/dfreproject/reproject.py
index 278d5e0..50dcd5a 100644
--- a/src/dfreproject/reproject.py
+++ b/src/dfreproject/reproject.py
@@ -702,8 +702,13 @@ class Reproject:
                 )
                 if self.compute_jacobian:  # Include Jacobian determinant computation
                     x_grid, y_grid = self.target_grid
-                    jacobian_det = self.compute_sip_jacobian(self.target_wcs, x_grid, y_grid)
-                    result[valid_pixels] = result[valid_pixels] * jacobian_det[valid_pixels]
+                    Jxx = x_source[..., :-1, 1:] - x_source[..., :-1, :-1]
+                    Jyx = y_source[..., :-1, 1:] - y_source[..., :-1, :-1]
+                    Jxy = x_source[..., 1:, :-1] - x_source[..., :-1, :-1]
+                    Jyy = y_source[..., 1:, :-1] - y_source[..., :-1, :-1]
+                    jacobian_det = Jxx * Jyy - Jxy * Jyx
+                    valid_pixels = valid_pixels[..., :-1, :-1]
+                    result[..., :-1, :-1][valid_pixels] = result[..., :-1, :-1][valid_pixels] * jacobian_det[valid_pixels[None, ...]]
             else:  # Do not apply flux conservation
                 result[valid_pixels] = combined_result[:, 0].squeeze()[valid_pixels]
         else:

```

Since we already have the corresponding input-image coordinate for each pixel in the output image, for Jxx I just subtracted `x_source` from `x_source` after offsetting it by one pixel in the x direction, to compute by finite differences how the input-image x coordinate changes if we shift by 1 pixel in output x coordinates, and likewise for the other Jacobian components. Then there's a lot of index work to account for the fact that this simple approach makes a Jacobian whose (x, y) shape is shortened by a pixel. A better implementation would probably either (a) use something like `np.gradient`, which will do centered finite-differences in the body of the array and un-centered at the edges, producing an output array the same shape as the input, or (b) make `x_source, y_source` larger by 1 in each direction (running from `(-1, shape)` rather than `(0, shape - 1)` in each dimension), and then use the full array to compute the Jacobian by centered finite differences but only use `x_source[1:-1]` for the interpolation.

(This issue part of openjournals/joss-reviews#8525)
