### 2025-09-04 DragonflyTelescope/dfreproject#27 @svank
And a smaller, related issue, in [test_flux_conservation](https://github.com/DragonflyTelescope/dfreproject/blob/03c746dab0de428e589359b70b982a42d9a23b20/tests/test_interpolation.py#L55C9-L56C27), I think the final comparison should be re-written as
```
err = abs(interpolated_flux - original_flux) / original_flux
assert err < 0.2
```
(the `1-` isn't needed on the `assert` line), and if the flux conservation is tightened as suggested, this test can look for something much tighter than agreement within 20%!
