### 2025-09-05 DragonflyTelescope/dfreproject#28 @svank
A few documentation question
There were a few spots where I wasn't sure about the documentation:

* [Here](https://dfreproject.readthedocs.io/en/latest/autoapi/dfreproject/reproject/index.html#dfreproject.reproject.Reproject.interpolate_source_image) and elsewhere where the flux conservation is discussed, the "ones" tensor wasn't clear to me at first, as the reprojected "ones" tensor should come out as all ones still. Is the idea that, for a reprojection like [this one](https://github.com/user-attachments/assets/80834606-68a1-4ecc-85e0-f4eb554dcea7), this makes a difference at the edge of where the input image falls in the output image---that the input image data only partially fills those output pixels, and so the "ones" tensor reduces those pixel values according to how much they're filled by the input image? If that's it (or if it's a different explanation), it may be helpful to explain the purpose of the "ones" tensor a little more in the documentation.
* Also in the `Reproject.interpolate_source_image` docstring, it says "Areas in the target image that map outside the source image boundaries will be filled with zeros (using ‘zeros’ padding_mode)." When I tested it I got `nan` in those pixels, not 0, and I don't see how to set `padding_mode`.
* In [this](https://dfreproject.readthedocs.io/en/latest/autoapi/dfreproject/reproject/index.html#dfreproject.reproject.calculate_reprojection) docstring, there's a formatting glitch, with `]` appearing as an argument.
* In the README, there is discussion of "flux density conservation", but I think it is now just "flux conservation".

(This issue part of openjournals/joss-reviews#8525)
