### 2025-09-05 DragonflyTelescope/dfreproject#29 @svank
Should tests cover tricky celestial coordinates?
I was looking through `test_coordinate_transformations.py`, as the coordinate transformations are critical, and I wonder if the tests could be improved by adding tests with different WCSes that include the celestial poles, and one with the dividing line between RA 360 and 0. I don't think those are covered, and those seem like spots where subtle bugs are mostly likely to lie. I think tests showing that `dfreproject` can correctly compute celestial coordinates for pixels right on those specific spots would increase this repo's robustness.

(This issue part of openjournals/joss-reviews#8525)
