### 2025-10-29 #8525 @crhea93
@warrickball Thank you! So for the paper, the authors listed are all the team members who worked with me to put together this project but they didn't necessarily contribute to the code -- hence them not being present in the archive. As for the authors in the archive, it looks like it is myself and Steve (who contributed code initially to the project) as well as the two reviews and a security bot. Should I remove them (Nathaniel, Adam, and the bot) from the archive and add the other authors on the paper?
