### 2025-10-30 6982b0c
Add autopublish to pypi with new tagged versions
