+++ .github/workflows/publish.yml
@@ -2,8 +2,10 @@ name: Publish to PyPI
 
 on:
   push:
+    branches:
+      - master
     tags:
-      - "v*" # Only run when pushing a tag like v1.0.0
+      - "v*" # Only run when a tag like v1.0.0 is pushed to master
 
 jobs:
   build-and-publish:
@@ -12,6 +14,8 @@ jobs:
     steps:
       - name: Check out source
         uses: actions/checkout@v4
+        with:
+          fetch-depth: 0 # Important: needed to fetch tags
 
       - name: Set up Python
         uses: actions/setup-python@v5
@@ -27,6 +31,7 @@ jobs:
         run: python -m build
 
       - name: Publish to PyPI
+        if: startsWith(github.ref, 'refs/tags/v')
         env:
           TWINE_USERNAME: __token__
           TWINE_PASSWORD: ${{ secrets.PYPI_API_TOKEN }}
