+++ .github/workflows/publish.yml
@@ -18,7 +18,7 @@ jobs:
           fetch-depth: 0 # Important: needed to fetch tags
 
       - name: Set up Python
-        uses: actions/setup-python@v5
+        uses: actions/setup-python@v6
         with:
           python-version: "3.11"
 
