### 2025-11-03 #8525 @warrickball
Sorry to iterate so much on these things but I'm trying to chip away given only a few slivers of time at the moment. I'll propose a few small editorial changes via PR in the next few days but in the meantime:

1. Could you please also change the archive *title* to match the paper? (Sorry I neglected to catch this the first time I checked the metadata.)
2. I also presumed you wanted the archive version to be v1.1.0, not v.1.1.0 (note the . after v), as in the Zenodo metadata, but I'm happy to update the version here to match.
3. Are the *David A. Dunlap Department of Astronomy & Astrophysics* and *Dunlap Institute for Astronomy & Astrophysics* (affiliations 4 and 5) actually different places? Or should they be one affiliation (AFAIK *Dunlap Institute for...*)?
