### 2025-11-04 #8525 @crhea93
Thank you for noticing that! I've done 1 and 2. As for 3, yes those are technically two separate affiliations.
