### 2025-11-04 08d0ac7
Initial editorial tweaks for JOSS publication
