### 2025-11-04 0177903
Update paper.md with captions
