### 2025-11-05 #8525 @warrickball
Congratulations @crhea93, your paper has been published in JOSS! Thanks for your patience through the final publication steps.

Thanks also and once again to @svank & @keflavich for reviewing and @adonath for editing this submission! JOSS simply wouldn't be possible without its community of volunteers.
